pyproject.toml
[project]
name = "assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==4.2.2",
    "chromadb>=0.5.5",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "google-api-core>=2.19.1",
    "google-genai>=0.3.0",
    "groq>=0.11.0",
    "ipython>=8.26.0",
    "jupyterlab>=4.5.5",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "numpy>=1.26.4",
    "openai>=2.26.0",
    "openai-whisper>=20250625",
    "pandas>=3.0.1",
    "pinecone>=8.1.0",
    "python-dotenv>=1.0.1",
    "pytube>=15.0.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.19.0",
    "torch>=2.10.0",
    "tqdm>=4.66.4",
    "transformers>=5.3.0",
    "watchdog>=6.0.0",
]
